move fips to its own file, and use a build tag
